Made a small refactorization of the code
